// examples/prisma-8-demo/src/prisma/contract.d.ts
export type CodecTypes = PgTypes & PgVectorTypes;
// ...
export type AddressOutput = {
  readonly city: CodecTypes['pg/text@1']['output'];
  readonly country: CodecTypes['pg/text@1']['output'];
  readonly street: CodecTypes['pg/text@1']['output'];
  readonly zip: CodecTypes['pg/text@1']['output'] | null;
};
// ...
export namespace Models {
  export type public_Bug = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    description: CodecTypes['pg/text@1']['output'] | null;
    id: CodecTypes['pg/uuid@1']['output'];
    status: CodecTypes['pg/text@1']['output'];
    title: CodecTypes['pg/text@1']['output'];
    type: 'bug';
    userId: CodecTypes['pg/uuid@1']['output'];
    severity: CodecTypes['pg/text@1']['output'];
    stepsToRepro: CodecTypes['pg/text@1']['output'] | null;
    user: public_User;
    readonly [RelationKeys]?: 'user';
  };
  export type public_Feature = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    description: CodecTypes['pg/text@1']['output'] | null;
    id: CodecTypes['pg/uuid@1']['output'];
    status: CodecTypes['pg/text@1']['output'];
    title: CodecTypes['pg/text@1']['output'];
    type: 'feature';
    userId: CodecTypes['pg/uuid@1']['output'];
    priority: CodecTypes['pg/text@1']['output'];
    targetRelease: CodecTypes['pg/text@1']['output'] | null;
    user: public_User;
    readonly [RelationKeys]?: 'user';
  };
  export type public_Post = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    embedding: Vector<1536> | null;
    expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    id: CodecTypes['pg/uuid@1']['output'];
    impressionCount: CodecTypes['pg/int8@1']['output'] | null;
    priority: 0 | 1 | 2;
    reachScore: CodecTypes['pg/unboundedint@1']['output'] | null;
    title: CodecTypes['pg/text@1']['output'];
    userId: CodecTypes['pg/uuid@1']['output'];
    viewCount: CodecTypes['pg/int8number@1']['output'] | null;
    tags: public_Tag[];
    user: public_User;
    readonly [RelationKeys]?: 'tags' | 'user';
  };
  export type public_PostTag = {
    postId: CodecTypes['pg/uuid@1']['output'];
    tagId: CodecTypes['pg/uuid@1']['output'];
    post: public_Post;
    tag: public_Tag;
    readonly [RelationKeys]?: 'post' | 'tag';
  };
  export type public_Tag = {
    id: CodecTypes['pg/uuid@1']['output'];
    label: CodecTypes['pg/text@1']['output'];
    posts: public_Post[];
    readonly [RelationKeys]?: 'posts';
  };
  export type public_Task = {
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    description: CodecTypes['pg/text@1']['output'] | null;
    id: CodecTypes['pg/uuid@1']['output'];
    status: CodecTypes['pg/text@1']['output'];
    title: CodecTypes['pg/text@1']['output'];
    type: 'bug' | 'feature';
    userId: CodecTypes['pg/uuid@1']['output'];
    user: public_User;
    readonly [RelationKeys]?: 'user';
  };
  export type public_User = {
    address: AddressOutput | null;
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    displayName: CodecTypes['pg/text@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/uuid@1']['output'];
    kind: 'admin' | 'user';
    posts: public_Post[];
    tasks: public_AnyTask[];
    readonly [RelationKeys]?: 'posts' | 'tasks';
  };
  export type public_AnyTask = public_Bug | public_Feature;
}